backend/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
Pillow
python-multipart
transformers
opencv-python
pandas
seaborn